# Situationsverwaltung/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.2"
  title: SitOPT API Reference
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - text/html

# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
...
  /things/ByName:
    x-swagger-router-controller: things
    get:
      tags:
        - thing
      summary: Get thing by name
      description: Returns array of things with specified name. Multiple things may have the same name.
      operationId: getThingByName
      
      parameters:
        - name: name
          in: query
          description: name of thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /things/ByID:
    x-swagger-router-controller: things
    get:
      tags:
        - thing
      summary: Get thing by ID
      description: Get thing by specified ID
      operationId: getThingByID
      
      parameters:
        - name: ID
          in: query
          description: ID of thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ThingModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    delete:
      tags:
        - thing
      summary: Delete thing by ID
      description: Deletes thing by specified ID
      operationId: deleteThingByID
      
      parameters:
        - name: ID
          in: query
          description: ID of thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /things:
  # binds a127 app logic to a route
    x-swagger-router-controller: things
    
    get:
      tags:
        - thing
      summary: Get all things
      description: Get all things
      operationId: allThings
      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/ThingModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
            
    post:
      tags:
        - thing
      summary: Save thing
      description: Save things in the database. Field "location" uses GeoJSON. Only use value of GeoJSON field "geometry"
      operationId: saveThing
      # used as the method name of the controller
      parameters:
        - in: body
          name: body
          description: input
          required: false
          schema:
            $ref: "#/definitions/ThingModel"
            
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /things/{ID}/{attribute}/{value}:
   # binds a127 app logic to a route
    x-swagger-router-controller: things
    
    post:
      tags:
        - thing
      summary: Update attribute
      description: Update or create attribute of Thing
      operationId: updateAttribute
      # used as the method name of the controller
      parameters:
        - name: "ID"
          in: "path"
          description: "ID of Thing"
          required: true
          type: "string"
        - name: "attribute"
          in: "path"
          description: "Name of attribute"
          required: true
          type: "string"
        - name: "value"
          in: "path"
          description: "Value of attribute"
          required: true
          type: "string"
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetModel:
    required:
      - name
    properties:
      name:
        type: string
  ThingModel:
    required:
      - name
      - sensors
      - owners
    properties:
      state:
        type: array
        items:
          type: string
      name:
        type: string
      url:
        type: string
      description:
        type: string
      sensors:
        type: array
        items:
          type: string
      location:
        type: string
      owners:
        type: array
        items:
          type: string
...
  Response:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string